Finished most of the form
Finished the majority of the form and the regular page styles. Need to finish form, form css, and comments for both. So far, for the form, I have used regular input  fields like for the name and phone number. I also created a plethora of selection lists; specifically for the monitor size, processor speed, et cetera. For the styles, I gave the HTML a background photo and created styles for the "nav" like giving it red lettering. I also changed the size of the banner logo.