Finished the rest of HTML and comments for it
Finished the yes/no input fields for the dvd burner option and more, the optional warranty checklist, and the comment box for users to request information/comments on their order. I also included three separate buttons at the end. They are to submit the order, cancel it, and to contact the buyer.